Move the package cmd routes into their own controller
Puts things together that belong together and makes SourceController smaller.
