+++ src/api/app/controllers/concerns/set_flag.rb
@@ -1,0 +1,22 @@
+module SetFlag
+  extend ActiveSupport::Concern
+
+  def obj_set_flag(obj)
+    obj.transaction do
+      begin
+        if params[:product]
+          obj.set_repository_by_product(params[:flag], params[:status], params[:product])
+        else
+          # first remove former flags of the same class
+          obj.remove_flag(params[:flag], params[:repository], params[:arch])
+          obj.add_flag(params[:flag], params[:status], params[:repository], params[:arch])
+        end
+      rescue ArgumentError => e
+        raise InvalidFlag, e.message
+      end
+
+      obj.store
+    end
+    render_ok
+  end
+end
